+++ src/main/php/webservices/rest/io/Buffered.class.php
@@ -7,7 +7,7 @@ class Buffered implements Transfer {
   public function writer($request, $payload, $format, $marshalling) {
     $bytes= $format->serialize($marshalling->marshal($payload), new MemoryOutputStream())->getBytes();
     $request->setHeader('Content-Length', strlen($bytes));
-    return function($request, $stream) use($bytes) {
+    return function($stream) use($bytes) {
       $stream->write($bytes);
       return $stream;
     };