+++ src/main/php/webservices/rest/Payload.class.php
@@ -1,0 +1,13 @@
+<?php namespace webservices\rest;
+
+class Payload {
+  private $value;
+
+  /** @param var $value */
+  public function __construct($value) {
+    $this->value= $value;
+  }
+
+  /** @return var */
+  public function value() { return $this->value; }
+}